[libc][stdlib] move abort to a header library and cleanup its usage
