+++ libc/src/compiler/generic/__stack_chk_fail.cpp
@@ -9,7 +9,7 @@
 #include "src/compiler/__stack_chk_fail.h"
 #include "hdr/stdint_proxy.h" // For uintptr_t
 #include "src/__support/OSUtil/io.h"
-#include "src/stdlib/abort.h"
+#include "src/stdlib/abort_utils.h"
 
 extern "C" {
 
@@ -17,7 +17,7 @@ uintptr_t __stack_chk_guard = static_cast<uintptr_t>(0xa9fff01234);
 
 void __stack_chk_fail(void) {
   LIBC_NAMESPACE::write_to_stderr("stack smashing detected\n");
-  LIBC_NAMESPACE::abort();
+  LIBC_NAMESPACE::abort_utils::abort();
 }
 
 } // extern "C"
