+++ libc/src/stdlib/baremetal/abort_utils.h
@@ -1,4 +1,4 @@
-//===-- Implementation of abort -------------------------------------------===//
+//===-- Internal header for baremetal abort ----------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,13 +6,18 @@
 //
 //===----------------------------------------------------------------------===//
 
+#ifndef LLVM_LIBC_SRC_STDLIB_BAREMETAL_ABORT_UTILS_H
+#define LLVM_LIBC_SRC_STDLIB_BAREMETAL_ABORT_UTILS_H
+
 #include "src/__support/common.h"
 #include "src/__support/macros/config.h"
 
-#include "src/stdlib/abort.h"
-
 namespace LIBC_NAMESPACE_DECL {
 
-LLVM_LIBC_FUNCTION(void, abort, ()) { __builtin_trap(); }
+namespace abort_utils {
+[[noreturn]] LIBC_INLINE void abort() { __builtin_trap(); }
+} // namespace abort_utils
 
 } // namespace LIBC_NAMESPACE_DECL
+
+#endif // LLVM_LIBC_SRC_STDLIB_BAREMETAL_ABORT_UTILS_H
