+++ libc/src/stdlib/gpu/abort_utils.h
@@ -1,4 +1,4 @@
-//===-- GPU implementation of abort ---------------------------------------===//
+//===-- Internal header for GPU abort ----------------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,16 +6,18 @@
 //
 //===----------------------------------------------------------------------===//
 
+#ifndef LLVM_LIBC_SRC_STDLIB_GPU_ABORT_UTILS_H
+#define LLVM_LIBC_SRC_STDLIB_GPU_ABORT_UTILS_H
+
 #include "src/__support/GPU/utils.h"
 #include "src/__support/RPC/rpc_client.h"
 #include "src/__support/common.h"
 #include "src/__support/macros/config.h"
 
-#include "src/stdlib/abort.h"
-
 namespace LIBC_NAMESPACE_DECL {
 
-LLVM_LIBC_FUNCTION(void, abort, ()) {
+namespace abort_utils {
+[[noreturn]] LIBC_INLINE void abort() {
   // We want to first make sure the server is listening before we abort.
   rpc::Client::Port port = rpc::client.open<LIBC_ABORT>();
   port.send_and_recv([](rpc::Buffer *, uint32_t) {},
@@ -24,5 +26,8 @@ LLVM_LIBC_FUNCTION(void, abort, ()) {
 
   gpu::end_program();
 }
+} // namespace abort_utils
 
 } // namespace LIBC_NAMESPACE_DECL
+
+#endif // LLVM_LIBC_SRC_STDLIB_GPU_ABORT_UTILS_H
